proper module dist
