+++ lib/slackdb.ex
@@ -7,8 +7,6 @@ defmodule SlackDB do
   (channel additions and archivals) made through this API will be updated
   and persisted in the supervisor channel
   """
-  use GenServer
-  use Private
   require Logger
   alias SlackDB.Client
   alias SlackDB.Utils
@@ -16,6 +14,7 @@ defmodule SlackDB do
   alias SlackDB.Search
   alias SlackDB.Key
   alias SlackDB.Channels
+  alias SlackDB.Server
 
   # @typedoc """
   # Types of SlackDB keys represented as atoms
@@ -43,27 +42,18 @@ defmodule SlackDB do
   """
   @type value :: String.t() | list(String.t())
 
-  @metadata_to_emoji %{
-    voting: ":thumbsup:",
-    multiple: ":family:",
-    single_front: ":hear_no_evil:",
-    single_back: ":monkey:",
-    constant: ":do_not_litter:",
-    undeletable: ":anchor:"
-    # locked: ":octagonal_sign:",
-  }
   ####################################################################################
   ## PUBLIC API ######################################################################
   ####################################################################################
 
-  @doc """
-  Start genserver, pull server config from application environment and pass to init/1
-  """
-  def start_link do
-    config = Application.get_env(:slackdb, :servers)
+  # @doc """
+  # Start genserver, pull server config from application environment and pass to init/1
+  # """
+  # def start_link do
+  #   config = Application.get_env(:slackdb, :servers)
 
-    GenServer.start_link(__MODULE__, config, name: __MODULE__)
-  end
+  #   GenServer.start_link(Server, config, name: Server)
+  # end
 
   @doc """
   Create a key/value pair in the specified server and channel of the given type
@@ -114,7 +104,7 @@ defmodule SlackDB do
     |> case do
       nil ->
         GenServer.call(
-          __MODULE__,
+          Server,
           {:create, server_name, channel_name, key_phrase, values, [key_type | add_metadata]}
         )
 
@@ -307,7 +297,7 @@ defmodule SlackDB do
          {:ok, %{"name" => channel_name, "id" => channel_id}} <-
            Client.conversations_create(user_token, channel_name, is_private?),
          {:ok, new_state} <-
-           GenServer.call(__MODULE__, {:put_channel, server_name, channel_name, channel_id}),
+           GenServer.call(Server, {:put_channel, server_name, channel_name, channel_id}),
          [{:ok, _resp}] <-
            append(
              server_name,
@@ -346,7 +336,7 @@ defmodule SlackDB do
 
         %{"id" => channel_id} ->
           with {:ok, new_state} <-
-                 GenServer.call(__MODULE__, {:put_channel, server_name, channel_name, channel_id}),
+                 GenServer.call(Server, {:put_channel, server_name, channel_name, channel_id}),
                [{:ok, _resp}] <-
                  append(
                    server_name,
@@ -378,8 +368,8 @@ defmodule SlackDB do
   def archive_channel(server_name, channel_name) do
     with %{supervisor_channel_name: supervisor_channel_name} <-
            Application.get_env(:slackdb, :servers) |> Map.get(server_name),
-         {:ok, _resp} <- GenServer.call(__MODULE__, {:archive, server_name, channel_name}),
-         new_state <- GenServer.call(__MODULE__, {:dump}),
+         {:ok, _resp} <- GenServer.call(Server, {:archive, server_name, channel_name}),
+         new_state <- GenServer.call(Server, {:dump}),
          [{:ok, _resp}] <-
            append(
              server_name,
@@ -410,7 +400,7 @@ defmodule SlackDB do
 
   def invite_to_channel(server_name, channel_name, user_ids)
       when is_list(user_ids) and length(user_ids) < 30 and length(user_ids) > 0 do
-    GenServer.call(__MODULE__, {:invite, server_name, channel_name, user_ids})
+    GenServer.call(Server, {:invite, server_name, channel_name, user_ids})
   end
 
   @doc """
@@ -464,169 +454,6 @@ defmodule SlackDB do
   """
   @spec dump() :: map()
   def dump() do
-    GenServer.call(__MODULE__, {:dump})
-  end
-
-  ####################################################################################
-  ## GENSERVER CALLBACKS #############################################################
-  ####################################################################################
-  @impl true
-  def init(config) do
-    servers = Map.keys(config)
-
-    server_states =
-      servers
-      |> Enum.map(fn server_name ->
-        {server_name,
-         read(
-           server_name,
-           get_in(config, [server_name, :supervisor_channel_name]),
-           server_name
-         )}
-      end)
-      |> Enum.map(&initialize_supervisor_channel/1)
-
-    config =
-      config
-      |> populate_channels(server_states)
-
-    # |> IO.inspect()
-
-    {:ok, config}
-  end
-
-  @impl true
-  def handle_call(
-        {:create, server_name, channel_name, key_phrase, values, all_metadata},
-        _from,
-        state
-      )
-      when is_list(values) do
-    resp =
-      with %{bot_token: bot_token, channels: channels} <- state[server_name] do
-        case channels[channel_name] do
-          nil ->
-            {:error, "channel_name_not_in_database"}
-
-          channel_id ->
-            metadata_as_emojis =
-              all_metadata |> Enum.map(fn x -> @metadata_to_emoji[x] end) |> Enum.join()
-
-            with {:ok, %{"ts" => thread_ts}} <-
-                   Client.chat_postMessage(
-                     bot_token,
-                     key_phrase <> " #{metadata_as_emojis}",
-                     channel_id
-                   ) do
-              Messages.post_thread(bot_token, channel_id, values, thread_ts)
-            else
-              err -> err
-            end
-        end
-      else
-        nil -> {:error, "server_not_found_in_config"}
-        %{} -> {:error, "improper_config"}
-        err -> err
-      end
-
-    {:reply, resp, state}
-  end
-
-  def handle_call({:invite, server_name, channel_name, user_ids}, _from, state)
-      when is_list(user_ids) do
-    resp =
-      with %{user_token: user_token, channels: channels} <- state[server_name] do
-        case channels[channel_name] do
-          nil ->
-            {:error, "channel_name_not_in_database"}
-
-          channel_id ->
-            Client.conversations_invite(user_token, channel_id, Enum.join(user_ids, ","))
-        end
-      else
-        nil -> {:error, "server_not_found_in_config"}
-        %{} -> {:error, "improper_config"}
-        err -> err
-      end
-
-    {:reply, resp, state}
-  end
-
-  def handle_call({:put_channel, server_name, channel_name, channel_id}, _from, state) do
-    new_state = state |> Utils.put_kv_in([server_name, :channels], channel_name, channel_id)
-    {:reply, {:ok, new_state}, new_state}
-  end
-
-  def handle_call({:archive, server_name, channel_name}, _from, state) do
-    resp =
-      with %{user_token: user_token, channels: channels} <- state[server_name] do
-        case channels[channel_name] do
-          nil ->
-            {:error, "channel_name_not_in_database"}
-
-          channel_id ->
-            Client.conversations_archive(user_token, channel_id)
-        end
-      else
-        nil -> {:error, "server_not_found_in_config"}
-        %{} -> {:error, "improper_config"}
-        err -> err
-      end
-
-    new_state =
-      case resp do
-        {:ok, _} ->
-          {_val, new} = state |> pop_in([server_name, :channels, channel_name])
-          new
-
-        {:error, _} ->
-          state
-      end
-
-    {:reply, resp, new_state}
-  end
-
-  def handle_call({:dump}, _from, state), do: {:reply, state, state}
-
-  ####################################################################################
-  ## INIT HELPERS ####################################################################
-  ####################################################################################
-
-  private do
-    defp populate_channels(map, [{server_name, state}]) do
-      Utils.put_kv_in(map, [server_name], :channels, state)
-    end
-
-    defp populate_channels(map, [{server_name, state} | more_server_states]) do
-      Utils.put_kv_in(populate_channels(map, more_server_states), [server_name], :channels, state)
-    end
-
-    defp initialize_supervisor_channel({server_name, {:ok, server_state}}) do
-      case server_state |> Jason.decode() do
-        {:ok, map} -> {server_name, map}
-        other -> {server_name, other}
-      end
-    end
-
-    defp initialize_supervisor_channel({server_name, {:error, _error_message}}) do
-      init =
-        with %{bot_token: bot_token, supervisor_channel_id: channel_id} <-
-               Application.get_env(:slackdb, :servers) |> Map.get(server_name),
-             {:ok, %{"ts" => thread_ts}} <-
-               Client.chat_postMessage(
-                 bot_token,
-                 server_name <> " #{@metadata_to_emoji[:single_back]}",
-                 channel_id
-               ),
-             _resp_list <- Messages.post_thread(bot_token, channel_id, "{}", thread_ts) do
-          %{}
-        else
-          nil -> {:error, "server_not_found_in_config"}
-          %{} -> {:error, "improper_config"}
-          err -> err
-        end
-
-      {server_name, init}
-    end
+    GenServer.call(Server, {:dump})
   end
 end
