+++ mix.exs
@@ -4,7 +4,7 @@ defmodule SlackDB.MixProject do
   def project do
     [
       app: :slackdb,
-      version: "0.1.1",
+      version: "0.2.0",
       elixir: "~> 1.7",
       start_permanent: Mix.env() == :prod,
       description: description(),
@@ -46,7 +46,9 @@ defmodule SlackDB.MixProject do
       {:jason, ">= 1.0.0"},
       {:private, "~> 0.1.1"},
       {:flow, "~> 0.14"},
-      {:ex_doc, ">= 0.0.0", only: :dev}
+      {:ex_doc, ">= 0.0.0", only: :dev},
+      {:httpoison, "~> 0.13.0"},
+      {:mox, "~> 0.5", only: :test}
     ]
   end
